renamed directories to be singular; removed remaining references to the visitors within the logger and basic sink
