+++ C++/Logger/conversion/ConvertibleTo.h
