+++ C++/Logger/conversion/Stringifiable.cpp
