+++ C++/Logger/conversion/Stringifiable.h
