+++ C++/Logger/conversion/VariantVisitor.cpp
