+++ C++/Logger/sink/BasicSink.cpp
@@ -1,11 +1,4 @@
 template <typename OutputType>
-BasicSink<OutputType>::BasicSink (void)
-: _outputVisitor ()
-{
-
-}
-
-template <typename OutputType>
 BasicSink<OutputType> & BasicSink<OutputType>::operator << (const Output & output)
 {
   this->output (OutputVariant::get (output));
